fixed int cast

+++ stratego/components/gameboard/GameScene.java
@@ -616,10 +616,10 @@ public class GameScene extends Pane {
 			temp = temp | (player1[x++] << 8 & 0x0000ff00);
 			temp = temp | (player1[x++] & 0x000000ff);
 			temp = Math.abs(temp-9);
-			
+
 			p1Arr[i].getRec().setY(((temp*72)*hFactor)+(startY*hFactor));
 		}
-		
+
 		x=0;
 		for(int i=0; i<40; i++){
 			temp = player2[x++] << 24 & 0xff000000;
@@ -643,12 +643,12 @@ public class GameScene extends Pane {
 		int y = 4;
 		int temp;
 		for (int i = 0; i < 40; i++) {
-			temp = (int)(Math.abs(p2Arr[i].getX()));
+			temp = (Math.abs(p2Arr[i].getX()));
 			p1[x++] = (byte) ((temp >> 24) & 0xff);
 			p1[x++] = (byte) ((temp >> 16) & 0xff);
 			p1[x++] = (byte) ((temp >> 8) & 0xff);
 			p1[x++] = (byte) temp;
-			temp = (int)(Math.abs(p2Arr[i].getY()));
+			temp = (Math.abs(p2Arr[i].getY()));
 			p1[y++] = (byte) ((temp >> 24) & 0xff);
 			p1[y++] = (byte) ((temp >> 16) & 0xff);
 			p1[y++] = (byte) ((temp >> 8) & 0xff);
@@ -665,12 +665,12 @@ public class GameScene extends Pane {
 		int y = 4;
 		int temp;
 		for (int i = 0; i < 40; i++) {
-			temp = (int)(Math.abs(p1Arr[i].getX()));
+			temp = (Math.abs(p1Arr[i].getX()));
 			p2[x++] = (byte) ((temp >> 24) & 0xff);
 			p2[x++] = (byte) ((temp >> 16) & 0xff);
 			p2[x++] = (byte) ((temp >> 8) & 0xff);
 			p2[x++] = (byte) temp;
-			temp = (int)(Math.abs(p1Arr[i].getY()));
+			temp = (Math.abs(p1Arr[i].getY()));
 			p2[y++] = (byte) ((temp >> 24) & 0xff);
 			p2[y++] = (byte) ((temp >> 16) & 0xff);
 			p2[y++] = (byte) ((temp >> 8) & 0xff);
